refactor: put new selectors in it's own module
