+++ app/components/UI/NetworkModal/index.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import NetworkModal from './index';
 import { render } from '@testing-library/react-native';
 import { useSelector } from 'react-redux';
-import { selectNetworkName } from '../../../selectors/networkController';
+import { selectNetworkName } from '../../../selectors/networkInfos';
 import { selectUseSafeChainsListValidation } from '../../../selectors/preferencesController';
 interface NetworkProps {
   isVisible: boolean;
